+++ src/lib.rs
@@ -164,7 +164,7 @@
 //! ```
 //!
 //! Alternatively, if you expect to only have component implementing the trait for each entity,
-//! you can use the filter [`One`](crate::One). This has significantly better performance than iterating
+//! you can use the filter [`One`](One). This has significantly better performance than iterating
 //! over all trait impls.
 //!
 //! ```ignore
@@ -189,6 +189,21 @@
 //! # bevy::ecs::system::assert_is_system(show_tooltips);
 //! ```
 //!
+//! Trait queries support basic change detection filtration. So to get all the components that
+//! implement the target trait, and have also changed in some way since the last tick, you can:
+//! ```ignore
+//! fn show_tooltips(
+//!     tooltips: Query<ChangedAll<&dyn Tooltip>>
+//!     // ...
+//! ) {
+//!     for tooltip in &tooltips {
+//!         println!("changed tool tips: {}", tooltip.tooltip());
+//!     }
+//! }
+//! ```
+//!
+//! Similarly, there exist `ChangedOne`, `AddedOne`, and `AddedAll`.
+//!
 //! # Performance
 //!
 //! The performance of trait queries is quite competitive. Here are some benchmarks for simple cases:
@@ -199,25 +214,30 @@
 //! | 2 matches         | 17.501 µs      | -                 | 102.83 µs       |
 //! | 1-2 matches       | -              | 16.959 µs         | 82.179 µs       |
 //!
-
-use std::cell::UnsafeCell;
-
 use bevy::{
     ecs::{
-        component::{ComponentId, ComponentTicks, StorageType},
-        query::{QueryItem, ReadOnlyWorldQuery, WorldQuery},
-        storage::{ComponentSparseSet, SparseSets, Table},
+        component::{ComponentId, StorageType},
         world::World,
     },
     prelude::*,
-    ptr::{Ptr, PtrMut, ThinSlicePtr, UnsafeCellDeref},
+    ptr::{Ptr, PtrMut},
 };
-use change_detection::{Mut, Ticks};
 
 #[cfg(test)]
 mod tests;
 
+pub mod all;
 pub mod change_detection;
+pub mod multiple_read;
+pub mod multiple_write;
+pub mod one;
+pub(crate) mod singular_read;
+pub(crate) mod singular_write;
+
+pub use all::*;
+pub use multiple_read::*;
+pub use multiple_write::*;
+pub use one::*;
 
 /// Marker for traits that can be used in queries.
 pub trait TraitQuery: 'static {}
@@ -331,6 +351,7 @@ impl<Trait: ?Sized + TraitQuery> TraitImplRegistry<Trait> {
             }
         }
     }
+
     fn seal(&mut self) {
         self.sealed = true;
     }
@@ -355,7 +376,9 @@ pub mod imports {
         archetype::{Archetype, ArchetypeComponentId},
         component::{Component, ComponentId},
         entity::Entity,
-        query::{Access, FilteredAccess, QueryItem, ReadOnlyWorldQuery, WorldQuery},
+        query::{
+            Access, Added, Changed, FilteredAccess, QueryItem, ReadOnlyWorldQuery, WorldQuery,
+        },
         storage::Table,
         world::World,
     };
@@ -385,10 +408,12 @@ impl<Trait: ?Sized + TraitQuery> TraitQueryState<Trait> {
             meta: registry.meta.clone().into_boxed_slice(),
         }
     }
+
     #[inline]
     fn matches_component_set_any(&self, set_contains_id: &impl Fn(ComponentId) -> bool) -> bool {
         self.components.iter().copied().any(set_contains_id)
     }
+
     #[inline]
     fn matches_component_set_one(&self, set_contains_id: &impl Fn(ComponentId) -> bool) -> bool {
         let match_count = self
@@ -458,1029 +483,6 @@ where
     let b = b.into_iter();
     debug_assert_eq!(a.len(), b.len());
     ZipExact { a, b }
-}
-
-/// [`WorldQuery`] adapter that fetches entities with exactly one component implementing a trait.
-pub struct One<T>(pub T);
-
-pub struct ReadTraitFetch<'w, Trait: ?Sized> {
-    // While we have shared access to all sparse set components,
-    // in practice we will only read the components specified in the `FetchState`.
-    // These accesses have been registered, which prevents runtime conflicts.
-    sparse_sets: &'w SparseSets,
-    // After `Fetch::set_archetype` or `set_table` has been called,
-    // this will carry the component data and metadata for the first trait impl found in the archetype.
-    storage: ReadStorage<'w, Trait>,
-}
-
-enum ReadStorage<'w, Trait: ?Sized> {
-    Uninit,
-    Table {
-        /// This points to one of the component table columns,
-        /// corresponding to one of the `ComponentId`s in the fetch state.
-        /// The fetch impl registers read access for all of these components,
-        /// so there will be no runtime conflicts.
-        column: Ptr<'w>,
-        meta: TraitImplMeta<Trait>,
-    },
-    SparseSet {
-        /// This gives us access to one of the components implementing the trait.
-        /// The fetch impl registers read access for all components implementing the trait,
-        /// so there will not be any runtime conflicts.
-        components: &'w ComponentSparseSet,
-        meta: TraitImplMeta<Trait>,
-    },
-}
-
-unsafe impl<'a, T: ?Sized + TraitQuery> ReadOnlyWorldQuery for One<&'a T> {}
-
-/// SAFETY: We only access the components registered in `DynQueryState`.
-/// This same set of components is used to match archetypes, and used to register world access.
-unsafe impl<'a, Trait: ?Sized + TraitQuery> WorldQuery for One<&'a Trait> {
-    type Item<'w> = &'w Trait;
-    type Fetch<'w> = ReadTraitFetch<'w, Trait>;
-    type ReadOnly = Self;
-    type State = TraitQueryState<Trait>;
-
-    #[inline]
-    fn shrink<'wlong: 'wshort, 'wshort>(item: QueryItem<'wlong, Self>) -> QueryItem<'wshort, Self> {
-        item
-    }
-
-    #[inline]
-    unsafe fn init_fetch<'w>(
-        world: &'w World,
-        _state: &Self::State,
-        _last_change_tick: u32,
-        _change_tick: u32,
-    ) -> ReadTraitFetch<'w, Trait> {
-        ReadTraitFetch {
-            storage: ReadStorage::Uninit,
-            sparse_sets: &world.storages().sparse_sets,
-        }
-    }
-
-    #[inline]
-    unsafe fn clone_fetch<'w>(fetch: &Self::Fetch<'w>) -> Self::Fetch<'w> {
-        ReadTraitFetch {
-            storage: match fetch.storage {
-                ReadStorage::Uninit => ReadStorage::Uninit,
-                ReadStorage::Table { column, meta } => ReadStorage::Table { column, meta },
-                ReadStorage::SparseSet { components, meta } => {
-                    ReadStorage::SparseSet { components, meta }
-                }
-            },
-            sparse_sets: fetch.sparse_sets,
-        }
-    }
-
-    const IS_DENSE: bool = false;
-    const IS_ARCHETYPAL: bool = false;
-
-    #[inline]
-    unsafe fn set_archetype<'w>(
-        fetch: &mut ReadTraitFetch<'w, Trait>,
-        state: &Self::State,
-        _archetype: &'w bevy::ecs::archetype::Archetype,
-        table: &'w bevy::ecs::storage::Table,
-    ) {
-        // Search for a registered trait impl that is present in the archetype.
-        // We check the table components first since it is faster to retrieve data of this type.
-        for (&component, &meta) in zip_exact(&*state.components, &*state.meta) {
-            if let Some(column) = table.get_column(component) {
-                fetch.storage = ReadStorage::Table {
-                    column: column.get_data_ptr(),
-                    meta,
-                };
-                return;
-            }
-        }
-        for (&component, &meta) in zip_exact(&*state.components, &*state.meta) {
-            if let Some(sparse_set) = fetch.sparse_sets.get(component) {
-                fetch.storage = ReadStorage::SparseSet {
-                    components: sparse_set,
-                    meta,
-                };
-                return;
-            }
-        }
-        // At least one of the components must be present in the table/sparse set.
-        debug_unreachable()
-    }
-
-    #[inline]
-    unsafe fn set_table<'w>(
-        fetch: &mut ReadTraitFetch<'w, Trait>,
-        state: &Self::State,
-        table: &'w bevy::ecs::storage::Table,
-    ) {
-        // Search for a registered trait impl that is present in the table.
-        for (&component, &meta) in std::iter::zip(&*state.components, &*state.meta) {
-            if let Some(column) = table.get_column(component) {
-                fetch.storage = ReadStorage::Table {
-                    column: column.get_data_ptr(),
-                    meta,
-                };
-                return;
-            }
-        }
-        // At least one of the components must be present in the table.
-        debug_unreachable()
-    }
-
-    #[inline]
-    unsafe fn fetch<'w>(
-        fetch: &mut Self::Fetch<'w>,
-        entity: Entity,
-        table_row: usize,
-    ) -> Self::Item<'w> {
-        match fetch.storage {
-            // SAFETY: This function must have been called after `set_archetype`,
-            // so we know that `self.storage` has been initialized.
-            ReadStorage::Uninit => debug_unreachable(),
-            ReadStorage::Table { column, meta } => {
-                let ptr = column.byte_add(table_row * meta.size_bytes);
-                meta.dyn_ctor.cast(ptr)
-            }
-            ReadStorage::SparseSet { components, meta } => {
-                let ptr = components
-                    .get(entity)
-                    .unwrap_or_else(|| debug_unreachable());
-                meta.dyn_ctor.cast(ptr)
-            }
-        }
-    }
-
-    #[inline]
-    fn update_component_access(
-        state: &Self::State,
-        access: &mut bevy::ecs::query::FilteredAccess<ComponentId>,
-    ) {
-        for &component in &*state.components {
-            assert!(
-                !access.access().has_write(component),
-                "&{} conflicts with a previous access in this query. Shared access cannot coincide with exclusive access.",
-                    std::any::type_name::<Trait>(),
-            );
-            access.add_read(component);
-        }
-    }
-
-    #[inline]
-    fn update_archetype_component_access(
-        state: &Self::State,
-        archetype: &bevy::ecs::archetype::Archetype,
-        access: &mut bevy::ecs::query::Access<bevy::ecs::archetype::ArchetypeComponentId>,
-    ) {
-        for &component in &*state.components {
-            if let Some(archetype_component_id) = archetype.get_archetype_component_id(component) {
-                access.add_read(archetype_component_id);
-            }
-        }
-    }
-
-    #[inline]
-    fn init_state(world: &mut World) -> Self::State {
-        TraitQueryState::init(world)
-    }
-    #[inline]
-    fn matches_component_set(
-        state: &Self::State,
-        set_contains_id: &impl Fn(ComponentId) -> bool,
-    ) -> bool {
-        state.matches_component_set_one(set_contains_id)
-    }
-}
-
-#[doc(hidden)]
-pub struct WriteTraitFetch<'w, Trait: ?Sized> {
-    // While we have shared mutable access to all sparse set components,
-    // in practice we will only modify the components specified in the `FetchState`.
-    // These accesses have been registered, which prevents runtime conflicts.
-    sparse_sets: &'w SparseSets,
-
-    // After `Fetch::set_archetype` or `set_table` has been called,
-    // this will carry the component data and metadata for the first trait impl found in the archetype.
-    storage: WriteStorage<'w, Trait>,
-
-    last_change_tick: u32,
-    change_tick: u32,
-}
-
-enum WriteStorage<'w, Trait: ?Sized> {
-    Uninit,
-    Table {
-        /// This is a shared mutable pointer to one of the component table columns,
-        /// corresponding to one of the `ComponentId`s in the fetch state.
-        /// The fetch impl registers write access for all of these components,
-        /// so there will be no runtime conflicts.
-        column: Ptr<'w>,
-        table_ticks: ThinSlicePtr<'w, UnsafeCell<ComponentTicks>>,
-        meta: TraitImplMeta<Trait>,
-    },
-    SparseSet {
-        /// This gives us shared mutable access to one of the components implementing the trait.
-        /// The fetch impl registers write access for all components implementing the trait, so there will be no runtime conflicts.
-        components: &'w ComponentSparseSet,
-        meta: TraitImplMeta<Trait>,
-    },
-}
-
-/// SAFETY: We only access the components registered in `DynQueryState`.
-/// This same set of components is used to match archetypes, and used to register world access.
-unsafe impl<'a, Trait: ?Sized + TraitQuery> WorldQuery for One<&'a mut Trait> {
-    type Item<'w> = Mut<'w, Trait>;
-    type Fetch<'w> = WriteTraitFetch<'w, Trait>;
-    type ReadOnly = One<&'a Trait>;
-    type State = TraitQueryState<Trait>;
-
-    #[inline]
-    unsafe fn init_fetch<'w>(
-        world: &'w World,
-        _state: &Self::State,
-        last_change_tick: u32,
-        change_tick: u32,
-    ) -> WriteTraitFetch<'w, Trait> {
-        WriteTraitFetch {
-            storage: WriteStorage::Uninit,
-            sparse_sets: &world.storages().sparse_sets,
-            last_change_tick,
-            change_tick,
-        }
-    }
-
-    #[inline]
-    unsafe fn clone_fetch<'w>(fetch: &Self::Fetch<'w>) -> Self::Fetch<'w> {
-        WriteTraitFetch {
-            storage: match fetch.storage {
-                WriteStorage::Uninit => WriteStorage::Uninit,
-                WriteStorage::Table {
-                    column,
-                    meta,
-                    table_ticks,
-                } => WriteStorage::Table {
-                    column,
-                    meta,
-                    table_ticks,
-                },
-                WriteStorage::SparseSet { components, meta } => {
-                    WriteStorage::SparseSet { components, meta }
-                }
-            },
-            sparse_sets: fetch.sparse_sets,
-            last_change_tick: fetch.last_change_tick,
-            change_tick: fetch.change_tick,
-        }
-    }
-
-    #[inline]
-    fn shrink<'wlong: 'wshort, 'wshort>(item: QueryItem<'wlong, Self>) -> QueryItem<'wshort, Self> {
-        item
-    }
-
-    const IS_DENSE: bool = false;
-    const IS_ARCHETYPAL: bool = false;
-
-    #[inline]
-    unsafe fn set_archetype<'w>(
-        fetch: &mut WriteTraitFetch<'w, Trait>,
-        state: &Self::State,
-        _archetype: &'w bevy::ecs::archetype::Archetype,
-        table: &'w bevy::ecs::storage::Table,
-    ) {
-        // Search for a registered trait impl that is present in the archetype.
-        for (&component, &meta) in zip_exact(&*state.components, &*state.meta) {
-            if let Some(column) = table.get_column(component) {
-                fetch.storage = WriteStorage::Table {
-                    column: column.get_data_ptr(),
-                    table_ticks: column.get_ticks_slice().into(),
-                    meta,
-                };
-                return;
-            }
-        }
-        for (&component, &meta) in zip_exact(&*state.components, &*state.meta) {
-            if let Some(sparse_set) = fetch.sparse_sets.get(component) {
-                fetch.storage = WriteStorage::SparseSet {
-                    components: sparse_set,
-                    meta,
-                };
-                return;
-            }
-        }
-        // At least one of the components must be present in the table/sparse set.
-        debug_unreachable()
-    }
-
-    #[inline]
-    unsafe fn set_table<'w>(
-        fetch: &mut WriteTraitFetch<'w, Trait>,
-        state: &Self::State,
-        table: &'w bevy::ecs::storage::Table,
-    ) {
-        // Search for a registered trait impl that is present in the table.
-        for (&component, &meta) in std::iter::zip(&*state.components, &*state.meta) {
-            if let Some(column) = table.get_column(component) {
-                fetch.storage = WriteStorage::Table {
-                    column: column.get_data_ptr(),
-                    table_ticks: column.get_ticks_slice().into(),
-                    meta,
-                };
-                return;
-            }
-        }
-        // At least one of the components must be present in the table.
-        debug_unreachable()
-    }
-
-    #[inline]
-    unsafe fn fetch<'w>(
-        fetch: &mut Self::Fetch<'w>,
-        entity: Entity,
-        table_row: usize,
-    ) -> Mut<'w, Trait> {
-        let dyn_ctor;
-        let (ptr, component_ticks) = match fetch.storage {
-            // SAFETY: This function must have been called after `set_archetype`,
-            // so we know that `self.storage` has been initialized.
-            WriteStorage::Uninit => debug_unreachable(),
-            WriteStorage::Table {
-                column,
-                table_ticks,
-                meta,
-            } => {
-                dyn_ctor = meta.dyn_ctor;
-                let ptr = column.byte_add(table_row * meta.size_bytes);
-                (
-                    // SAFETY: `column` allows for shared mutable access.
-                    // So long as the caller does not invoke this function twice with the same archetype_index,
-                    // this pointer will never be aliased.
-                    ptr.assert_unique(),
-                    // SAFETY: We have exclusive access to the component, so by extension
-                    // we have exclusive access to the corresponding `ComponentTicks`.
-                    table_ticks.get(table_row).deref_mut(),
-                )
-            }
-            WriteStorage::SparseSet { components, meta } => {
-                dyn_ctor = meta.dyn_ctor;
-                let (ptr, ticks) = components
-                    .get_with_ticks(entity)
-                    .unwrap_or_else(|| debug_unreachable());
-                (
-                    // SAFETY: We have exclusive access to the sparse set `components`.
-                    // So long as the caller does not invoke this function twice with the same archetype_index,
-                    // this pointer will never be aliased.
-                    ptr.assert_unique(),
-                    // SAFETY: We have exclusive access to the component, so by extension
-                    // we have exclusive access to the corresponding `ComponentTicks`.
-                    ticks.deref_mut(),
-                )
-            }
-        };
-
-        Mut {
-            value: dyn_ctor.cast_mut(ptr),
-            ticks: Ticks {
-                component_ticks,
-                last_change_tick: fetch.last_change_tick,
-                change_tick: fetch.change_tick,
-            },
-        }
-    }
-
-    #[inline]
-    fn update_component_access(
-        state: &Self::State,
-        access: &mut bevy::ecs::query::FilteredAccess<ComponentId>,
-    ) {
-        for &component in &*state.components {
-            assert!(
-                !access.access().has_write(component),
-                "&mut {} conflicts with a previous access in this query. Mutable component access must be unique.",
-                    std::any::type_name::<Trait>(),
-            );
-            access.add_write(component);
-        }
-    }
-
-    #[inline]
-    fn update_archetype_component_access(
-        state: &Self::State,
-        archetype: &bevy::ecs::archetype::Archetype,
-        access: &mut bevy::ecs::query::Access<bevy::ecs::archetype::ArchetypeComponentId>,
-    ) {
-        for &component in &*state.components {
-            if let Some(archetype_component_id) = archetype.get_archetype_component_id(component) {
-                access.add_write(archetype_component_id);
-            }
-        }
-    }
-
-    #[inline]
-    fn init_state(world: &mut World) -> Self::State {
-        TraitQueryState::init(world)
-    }
-    #[inline]
-    fn matches_component_set(
-        state: &Self::State,
-        set_contains_id: &impl Fn(ComponentId) -> bool,
-    ) -> bool {
-        state.matches_component_set_one(set_contains_id)
-    }
-}
-
-/// `WorldQuery` adapter that fetches all implementations of a given trait for an entity.
-///
-/// You can usually just use `&dyn Trait` or `&mut dyn Trait` as a `WorldQuery` directly.
-pub struct All<T: ?Sized>(T);
-
-/// Read-access to all components implementing a trait for a given entity.
-pub struct ReadTraits<'a, Trait: ?Sized + TraitQuery> {
-    // Read-only access to the global trait registry.
-    // Since no one outside of the crate can name the registry type,
-    // we can be confident that no write accessess will conflict with this.
-    registry: &'a TraitImplRegistry<Trait>,
-
-    table: &'a Table,
-    table_row: usize,
-
-    /// This grants shared access to all sparse set components,
-    /// but in practice we will only read the components specified in `self.registry`.
-    /// The fetch impl registers read-access for all of these components,
-    /// so there will be no runtime conflicts.
-    sparse_sets: &'a SparseSets,
-}
-
-/// Write-access to all components implementing a trait for a given entity.
-pub struct WriteTraits<'a, Trait: ?Sized + TraitQuery> {
-    // Read-only access to the global trait registry.
-    // Since no one outside of the crate can name the registry type,
-    // we can be confident that no write accessess will conflict with this.
-    registry: &'a TraitImplRegistry<Trait>,
-
-    table: &'a Table,
-    table_row: usize,
-
-    last_change_tick: u32,
-    change_tick: u32,
-
-    /// This grants shared mutable access to all sparse set components,
-    /// but in practice we will only modify the components specified in `self.registry`.
-    /// The fetch impl registers write-access for all of these components,
-    /// guaranteeing us exclusive access at runtime.
-    sparse_sets: &'a SparseSets,
-}
-
-#[doc(hidden)]
-pub type CombinedReadTraitsIter<'a, Trait> =
-    std::iter::Chain<ReadTableTraitsIter<'a, Trait>, ReadSparseTraitsIter<'a, Trait>>;
-
-#[doc(hidden)]
-pub type CombinedWriteTraitsIter<'a, Trait> =
-    std::iter::Chain<WriteTableTraitsIter<'a, Trait>, WriteSparseTraitsIter<'a, Trait>>;
-
-#[doc(hidden)]
-pub struct ReadTableTraitsIter<'a, Trait: ?Sized> {
-    // SAFETY: These two iterators must have equal length.
-    components: std::slice::Iter<'a, ComponentId>,
-    meta: std::slice::Iter<'a, TraitImplMeta<Trait>>,
-    table_row: usize,
-    // Grants shared access to the components corresponding to `components` in this table.
-    // Not all components are guaranteed to exist in the table.
-    table: &'a Table,
-}
-
-impl<'a, Trait: ?Sized + TraitQuery> Iterator for ReadTableTraitsIter<'a, Trait> {
-    type Item = &'a Trait;
-    fn next(&mut self) -> Option<Self::Item> {
-        // Iterate the remaining table components that are registered,
-        // until we find one that exists in the table.
-        let (column, meta) = unsafe { zip_exact(&mut self.components, &mut self.meta) }
-            .find_map(|(&component, meta)| self.table.get_column(component).zip(Some(meta)))?;
-        // SAFETY: We have shared access to the entire column.
-        let ptr = unsafe {
-            column
-                .get_data_ptr()
-                .byte_add(self.table_row * meta.size_bytes)
-        };
-        let trait_object = unsafe { meta.dyn_ctor.cast(ptr) };
-        Some(trait_object)
-    }
-}
-
-#[doc(hidden)]
-pub struct ReadSparseTraitsIter<'a, Trait: ?Sized> {
-    // SAFETY: These two iterators must have equal length.
-    components: std::slice::Iter<'a, ComponentId>,
-    meta: std::slice::Iter<'a, TraitImplMeta<Trait>>,
-    entity: Entity,
-    // Grants shared access to the components corresponding to both `components` and `entity`.
-    sparse_sets: &'a SparseSets,
-}
-
-impl<'a, Trait: ?Sized + TraitQuery> Iterator for ReadSparseTraitsIter<'a, Trait> {
-    type Item = &'a Trait;
-    fn next(&mut self) -> Option<Self::Item> {
-        // Iterate the remaining sparse set components that are registered,
-        // until we find one that exists in the archetype.
-        let (ptr, meta) = unsafe { zip_exact(&mut self.components, &mut self.meta) }.find_map(
-            |(&component, meta)| {
-                self.sparse_sets
-                    .get(component)
-                    .and_then(|set| set.get(self.entity))
-                    .zip(Some(meta))
-            },
-        )?;
-        let trait_object = unsafe { meta.dyn_ctor.cast(ptr) };
-        Some(trait_object)
-    }
-}
-
-#[doc(hidden)]
-pub struct WriteTableTraitsIter<'a, Trait: ?Sized> {
-    // SAFETY: These two iterators must have equal length.
-    components: std::slice::Iter<'a, ComponentId>,
-    meta: std::slice::Iter<'a, TraitImplMeta<Trait>>,
-    table: &'a Table,
-    /// SAFETY: Given the same trait type and same archetype,
-    /// no two instances of this struct may have the same `table_row`.
-    table_row: usize,
-    last_change_tick: u32,
-    change_tick: u32,
-}
-
-impl<'a, Trait: ?Sized + TraitQuery> Iterator for WriteTableTraitsIter<'a, Trait> {
-    type Item = Mut<'a, Trait>;
-    fn next(&mut self) -> Option<Self::Item> {
-        // Iterate the remaining table components that are registered,
-        // until we find one that exists in the table.
-        let (column, meta) = unsafe { zip_exact(&mut self.components, &mut self.meta) }
-            .find_map(|(&component, meta)| self.table.get_column(component).zip(Some(meta)))?;
-        let ptr = unsafe {
-            column
-                .get_data_ptr()
-                .byte_add(self.table_row * meta.size_bytes)
-        };
-        // SAFETY: The instance of `WriteTraits` that created this iterator
-        // has exclusive access to all table components registered with the trait.
-        //
-        // Since `self.table_row` is guaranteed to be unique, we know that other instances
-        // of `WriteTableTraitsIter` will not conflict with this pointer.
-        let ptr = unsafe { ptr.assert_unique() };
-        let trait_object = unsafe { meta.dyn_ctor.cast_mut(ptr) };
-        // SAFETY: We have exclusive access to the component, so by extension
-        // we have exclusive access to the corresponding `ComponentTicks`.
-        let component_ticks = unsafe { column.get_ticks_unchecked(self.table_row).deref_mut() };
-        Some(Mut {
-            value: trait_object,
-            ticks: Ticks {
-                component_ticks,
-                last_change_tick: self.last_change_tick,
-                change_tick: self.change_tick,
-            },
-        })
-    }
-}
-
-#[doc(hidden)]
-pub struct WriteSparseTraitsIter<'a, Trait: ?Sized> {
-    // SAFETY: These two iterators must have equal length.
-    components: std::slice::Iter<'a, ComponentId>,
-    meta: std::slice::Iter<'a, TraitImplMeta<Trait>>,
-    /// SAFETY: Given the same trait type and same archetype,
-    /// no two instances of this struct may have the same `entity`.
-    entity: Entity,
-    sparse_sets: &'a SparseSets,
-    last_change_tick: u32,
-    change_tick: u32,
-}
-
-impl<'a, Trait: ?Sized + TraitQuery> Iterator for WriteSparseTraitsIter<'a, Trait> {
-    type Item = Mut<'a, Trait>;
-    fn next(&mut self) -> Option<Self::Item> {
-        // Iterate the remaining sparse set components we have registered,
-        // until we find one that exists in the archetype.
-        let ((ptr, component_ticks), meta) =
-            unsafe { zip_exact(&mut self.components, &mut self.meta) }.find_map(
-                |(&component, meta)| {
-                    self.sparse_sets
-                        .get(component)
-                        .and_then(|set| set.get_with_ticks(self.entity))
-                        .zip(Some(meta))
-                },
-            )?;
-
-        // SAFETY: The instance of `WriteTraits` that created this iterator
-        // has exclusive access to all sparse set components registered with the trait.
-        //
-        // Since `self.entity` is guaranteed to be unique, we know that other instances
-        // of `WriteSparseTraitsIter` will not conflict with this pointer.
-        let ptr = unsafe { ptr.assert_unique() };
-        let trait_object = unsafe { meta.dyn_ctor.cast_mut(ptr) };
-        // SAFETY: We have exclusive access to the component, so by extension
-        // we have exclusive access to the corresponding `ComponentTicks`.
-        let component_ticks = unsafe { component_ticks.deref_mut() };
-
-        Some(Mut {
-            value: trait_object,
-            ticks: Ticks {
-                component_ticks,
-                last_change_tick: self.last_change_tick,
-                change_tick: self.change_tick,
-            },
-        })
-    }
-}
-
-#[doc(hidden)]
-pub struct ReadAllTraitsFetch<'w, Trait: ?Sized> {
-    registry: &'w TraitImplRegistry<Trait>,
-    table: Option<&'w Table>,
-    sparse_sets: &'w SparseSets,
-}
-
-#[doc(hidden)]
-pub struct WriteAllTraitsFetch<'w, Trait: ?Sized + TraitQuery> {
-    registry: &'w TraitImplRegistry<Trait>,
-    table: Option<&'w Table>,
-    sparse_sets: &'w SparseSets,
-
-    last_change_tick: u32,
-    change_tick: u32,
-}
-
-unsafe impl<'a, Trait: ?Sized + TraitQuery> ReadOnlyWorldQuery for All<&'a Trait> {}
-
-/// SAFETY: We only access the components registered in the trait registry.
-/// This is known to match the set of components in the `DynQueryState`,
-/// which is used to match archetypes and register world access.
-unsafe impl<'a, Trait: ?Sized + TraitQuery> WorldQuery for All<&'a Trait> {
-    type Item<'w> = ReadTraits<'w, Trait>;
-    type Fetch<'w> = ReadAllTraitsFetch<'w, Trait>;
-    type ReadOnly = Self;
-    type State = TraitQueryState<Trait>;
-
-    #[inline]
-    unsafe fn init_fetch<'w>(
-        world: &'w World,
-        _state: &Self::State,
-        _last_change_tick: u32,
-        _change_tick: u32,
-    ) -> ReadAllTraitsFetch<'w, Trait> {
-        ReadAllTraitsFetch {
-            registry: world.resource(),
-            table: None,
-            sparse_sets: &world.storages().sparse_sets,
-        }
-    }
-
-    #[inline]
-    unsafe fn clone_fetch<'w>(fetch: &Self::Fetch<'w>) -> Self::Fetch<'w> {
-        ReadAllTraitsFetch {
-            registry: fetch.registry,
-            table: fetch.table,
-            sparse_sets: fetch.sparse_sets,
-        }
-    }
-
-    #[inline]
-    fn shrink<'wlong: 'wshort, 'wshort>(item: QueryItem<'wlong, Self>) -> QueryItem<'wshort, Self> {
-        item
-    }
-
-    const IS_DENSE: bool = false;
-    const IS_ARCHETYPAL: bool = false;
-
-    #[inline]
-    unsafe fn set_archetype<'w>(
-        fetch: &mut ReadAllTraitsFetch<'w, Trait>,
-        _state: &Self::State,
-        _archetype: &'w bevy::ecs::archetype::Archetype,
-        table: &'w bevy::ecs::storage::Table,
-    ) {
-        fetch.table = Some(table);
-    }
-
-    unsafe fn set_table<'w>(
-        fetch: &mut ReadAllTraitsFetch<'w, Trait>,
-        _state: &Self::State,
-        table: &'w bevy::ecs::storage::Table,
-    ) {
-        fetch.table = Some(table);
-    }
-
-    #[inline]
-    unsafe fn fetch<'w>(
-        fetch: &mut Self::Fetch<'w>,
-        _entity: Entity,
-        table_row: usize,
-    ) -> Self::Item<'w> {
-        let table = fetch.table.unwrap_or_else(|| debug_unreachable());
-
-        ReadTraits {
-            registry: fetch.registry,
-            table,
-            table_row,
-            sparse_sets: fetch.sparse_sets,
-        }
-    }
-
-    #[inline]
-    fn update_component_access(
-        state: &Self::State,
-        access: &mut bevy::ecs::query::FilteredAccess<ComponentId>,
-    ) {
-        for &component in &*state.components {
-            assert!(
-                !access.access().has_write(component),
-                "&{} conflicts with a previous access in this query. Shared access cannot coincide with exclusive access.",
-                    std::any::type_name::<Trait>(),
-            );
-            access.add_read(component);
-        }
-    }
-
-    #[inline]
-    fn update_archetype_component_access(
-        state: &Self::State,
-        archetype: &bevy::ecs::archetype::Archetype,
-        access: &mut bevy::ecs::query::Access<bevy::ecs::archetype::ArchetypeComponentId>,
-    ) {
-        for &component in &*state.components {
-            if let Some(archetype_component_id) = archetype.get_archetype_component_id(component) {
-                access.add_read(archetype_component_id);
-            }
-        }
-    }
-
-    #[inline]
-    fn init_state(world: &mut World) -> Self::State {
-        TraitQueryState::init(world)
-    }
-    #[inline]
-    fn matches_component_set(
-        state: &Self::State,
-        set_contains_id: &impl Fn(ComponentId) -> bool,
-    ) -> bool {
-        state.matches_component_set_any(set_contains_id)
-    }
-}
-
-/// SAFETY: We only access the components registered in the trait registry.
-/// This is known to match the set of components in the `DynQueryState`,
-/// which is used to match archetypes and register world access.
-unsafe impl<'a, Trait: ?Sized + TraitQuery> WorldQuery for All<&'a mut Trait> {
-    type Item<'w> = WriteTraits<'w, Trait>;
-    type Fetch<'w> = WriteAllTraitsFetch<'w, Trait>;
-    type ReadOnly = All<&'a Trait>;
-    type State = TraitQueryState<Trait>;
-
-    #[inline]
-    unsafe fn init_fetch<'w>(
-        world: &'w World,
-        _state: &Self::State,
-        last_change_tick: u32,
-        change_tick: u32,
-    ) -> WriteAllTraitsFetch<'w, Trait> {
-        WriteAllTraitsFetch {
-            registry: world.resource(),
-            table: None,
-            sparse_sets: &world.storages().sparse_sets,
-            last_change_tick,
-            change_tick,
-        }
-    }
-
-    #[inline]
-    unsafe fn clone_fetch<'w>(fetch: &Self::Fetch<'w>) -> Self::Fetch<'w> {
-        WriteAllTraitsFetch {
-            registry: fetch.registry,
-            table: fetch.table,
-            sparse_sets: fetch.sparse_sets,
-            last_change_tick: fetch.last_change_tick,
-            change_tick: fetch.change_tick,
-        }
-    }
-
-    #[inline]
-    fn shrink<'wlong: 'wshort, 'wshort>(item: QueryItem<'wlong, Self>) -> QueryItem<'wshort, Self> {
-        item
-    }
-
-    const IS_DENSE: bool = false;
-    const IS_ARCHETYPAL: bool = false;
-
-    #[inline]
-    unsafe fn set_archetype<'w>(
-        fetch: &mut WriteAllTraitsFetch<'w, Trait>,
-        _state: &Self::State,
-        _archetype: &'w bevy::ecs::archetype::Archetype,
-        table: &'w bevy::ecs::storage::Table,
-    ) {
-        fetch.table = Some(table);
-    }
-
-    #[inline]
-    unsafe fn fetch<'w>(
-        fetch: &mut Self::Fetch<'w>,
-        _entity: Entity,
-        table_row: usize,
-    ) -> Self::Item<'w> {
-        let table = fetch.table.unwrap_or_else(|| debug_unreachable());
-
-        WriteTraits {
-            registry: fetch.registry,
-            table,
-            table_row,
-            sparse_sets: fetch.sparse_sets,
-            last_change_tick: fetch.last_change_tick,
-            change_tick: fetch.change_tick,
-        }
-    }
-
-    #[inline]
-    unsafe fn set_table<'w>(
-        fetch: &mut WriteAllTraitsFetch<'w, Trait>,
-        _state: &Self::State,
-        table: &'w bevy::ecs::storage::Table,
-    ) {
-        fetch.table = Some(table);
-    }
-
-    #[inline]
-    fn update_component_access(
-        state: &Self::State,
-        access: &mut bevy::ecs::query::FilteredAccess<ComponentId>,
-    ) {
-        for &component in &*state.components {
-            assert!(
-                !access.access().has_write(component),
-                "&mut {} conflicts with a previous access in this query. Mutable component access must be unique.",
-                    std::any::type_name::<Trait>(),
-            );
-            access.add_write(component);
-        }
-    }
-
-    #[inline]
-    fn update_archetype_component_access(
-        state: &Self::State,
-        archetype: &bevy::ecs::archetype::Archetype,
-        access: &mut bevy::ecs::query::Access<bevy::ecs::archetype::ArchetypeComponentId>,
-    ) {
-        for &component in &*state.components {
-            if let Some(archetype_component_id) = archetype.get_archetype_component_id(component) {
-                access.add_write(archetype_component_id);
-            }
-        }
-    }
-
-    #[inline]
-    fn init_state(world: &mut World) -> Self::State {
-        TraitQueryState::init(world)
-    }
-    #[inline]
-    fn matches_component_set(
-        state: &Self::State,
-        set_contains_id: &impl Fn(ComponentId) -> bool,
-    ) -> bool {
-        state.matches_component_set_any(set_contains_id)
-    }
-}
-
-impl<'w, Trait: ?Sized + TraitQuery> IntoIterator for ReadTraits<'w, Trait> {
-    type Item = &'w Trait;
-    type IntoIter = CombinedReadTraitsIter<'w, Trait>;
-    #[inline]
-    fn into_iter(self) -> Self::IntoIter {
-        let table = ReadTableTraitsIter {
-            components: self.registry.table_components.iter(),
-            meta: self.registry.table_meta.iter(),
-            table: self.table,
-            table_row: self.table_row,
-        };
-        let sparse = ReadSparseTraitsIter {
-            components: self.registry.sparse_components.iter(),
-            meta: self.registry.sparse_meta.iter(),
-            entity: self.table.entities()[self.table_row],
-            sparse_sets: self.sparse_sets,
-        };
-        table.chain(sparse)
-    }
-}
-
-impl<'w, Trait: ?Sized + TraitQuery> ReadTraits<'w, Trait> {
-    /// Returns an iterator over the components implementing `Trait` for the current entity.
-    pub fn iter(&self) -> CombinedReadTraitsIter<'w, Trait> {
-        self.into_iter()
-    }
-}
-
-impl<'w, Trait: ?Sized + TraitQuery> WriteTraits<'w, Trait> {
-    /// Returns an iterator over the components implementing `Trait` for the current entity.
-    pub fn iter(&self) -> CombinedReadTraitsIter<'_, Trait> {
-        self.into_iter()
-    }
-    /// Returns a mutable iterator over the components implementing `Trait` for the current entity.
-    pub fn iter_mut(&mut self) -> CombinedWriteTraitsIter<'_, Trait> {
-        self.into_iter()
-    }
-}
-
-impl<'w, Trait: ?Sized + TraitQuery> IntoIterator for &ReadTraits<'w, Trait> {
-    type Item = &'w Trait;
-    type IntoIter = CombinedReadTraitsIter<'w, Trait>;
-    #[inline]
-    fn into_iter(self) -> Self::IntoIter {
-        let table = ReadTableTraitsIter {
-            components: self.registry.table_components.iter(),
-            meta: self.registry.table_meta.iter(),
-            table: self.table,
-            table_row: self.table_row,
-        };
-        let sparse = ReadSparseTraitsIter {
-            components: self.registry.sparse_components.iter(),
-            meta: self.registry.sparse_meta.iter(),
-            entity: self.table.entities()[self.table_row],
-            sparse_sets: self.sparse_sets,
-        };
-        table.chain(sparse)
-    }
-}
-
-impl<'w, Trait: ?Sized + TraitQuery> IntoIterator for WriteTraits<'w, Trait> {
-    type Item = Mut<'w, Trait>;
-    type IntoIter = CombinedWriteTraitsIter<'w, Trait>;
-    #[inline]
-    fn into_iter(self) -> Self::IntoIter {
-        let table = WriteTableTraitsIter {
-            components: self.registry.table_components.iter(),
-            meta: self.registry.table_meta.iter(),
-            table: self.table,
-            table_row: self.table_row,
-            last_change_tick: self.last_change_tick,
-            change_tick: self.change_tick,
-        };
-        let sparse = WriteSparseTraitsIter {
-            components: self.registry.sparse_components.iter(),
-            meta: self.registry.sparse_meta.iter(),
-            entity: self.table.entities()[self.table_row],
-            sparse_sets: self.sparse_sets,
-            last_change_tick: self.last_change_tick,
-            change_tick: self.change_tick,
-        };
-        table.chain(sparse)
-    }
-}
-
-impl<'world, 'local, Trait: ?Sized + TraitQuery> IntoIterator
-    for &'local WriteTraits<'world, Trait>
-{
-    type Item = &'local Trait;
-    type IntoIter = CombinedReadTraitsIter<'local, Trait>;
-    #[inline]
-    fn into_iter(self) -> Self::IntoIter {
-        let table = ReadTableTraitsIter {
-            components: self.registry.table_components.iter(),
-            meta: self.registry.table_meta.iter(),
-            table: self.table,
-            table_row: self.table_row,
-        };
-        let sparse = ReadSparseTraitsIter {
-            components: self.registry.sparse_components.iter(),
-            meta: self.registry.sparse_meta.iter(),
-            entity: self.table.entities()[self.table_row],
-            sparse_sets: self.sparse_sets,
-        };
-        table.chain(sparse)
-    }
-}
-
-impl<'world, 'local, Trait: ?Sized + TraitQuery> IntoIterator
-    for &'local mut WriteTraits<'world, Trait>
-{
-    type Item = Mut<'local, Trait>;
-    type IntoIter = CombinedWriteTraitsIter<'local, Trait>;
-    #[inline]
-    fn into_iter(self) -> Self::IntoIter {
-        let table = WriteTableTraitsIter {
-            components: self.registry.table_components.iter(),
-            meta: self.registry.table_meta.iter(),
-            table: self.table,
-            table_row: self.table_row,
-            last_change_tick: self.last_change_tick,
-            change_tick: self.change_tick,
-        };
-        let sparse = WriteSparseTraitsIter {
-            components: self.registry.sparse_components.iter(),
-            meta: self.registry.sparse_meta.iter(),
-            entity: self.table.entities()[self.table_row],
-            sparse_sets: self.sparse_sets,
-            last_change_tick: self.last_change_tick,
-            change_tick: self.change_tick,
-        };
-        table.chain(sparse)
-    }
 }
 
 #[track_caller]
