+++ src/multiple_write.rs
@@ -1,0 +1,222 @@
+use crate::change_detection::{Mut, Ticks};
+use crate::multiple_read::{CombinedReadTraitsIter, ReadSparseTraitsIter, ReadTableTraitsIter};
+use crate::{zip_exact, TraitImplMeta, TraitImplRegistry, TraitQuery};
+use bevy::ecs::component::ComponentId;
+use bevy::ecs::entity::Entity;
+use bevy::ecs::storage::{SparseSets, Table};
+use bevy::ptr::UnsafeCellDeref;
+
+/// Write-access to all components implementing a trait for a given entity.
+pub struct WriteTraits<'a, Trait: ?Sized + TraitQuery> {
+    // Read-only access to the global trait registry.
+    // Since no one outside of the crate can name the registry type,
+    // we can be confident that no write accesses will conflict with this.
+    pub(crate) registry: &'a TraitImplRegistry<Trait>,
+
+    pub(crate) table: &'a Table,
+    pub(crate) table_row: usize,
+
+    pub(crate) last_change_tick: u32,
+    pub(crate) change_tick: u32,
+
+    /// This grants shared mutable access to all sparse set components,
+    /// but in practice we will only modify the components specified in `self.registry`.
+    /// The fetch impl registers write-access for all of these components,
+    /// guaranteeing us exclusive access at runtime.
+    pub(crate) sparse_sets: &'a SparseSets,
+}
+
+#[doc(hidden)]
+pub type CombinedWriteTraitsIter<'a, Trait> =
+    std::iter::Chain<WriteTableTraitsIter<'a, Trait>, WriteSparseTraitsIter<'a, Trait>>;
+
+#[doc(hidden)]
+pub struct WriteTableTraitsIter<'a, Trait: ?Sized> {
+    // SAFETY: These two iterators must have equal length.
+    components: std::slice::Iter<'a, ComponentId>,
+    meta: std::slice::Iter<'a, TraitImplMeta<Trait>>,
+    table: &'a Table,
+    /// SAFETY: Given the same trait type and same archetype,
+    /// no two instances of this struct may have the same `table_row`.
+    table_row: usize,
+    last_change_tick: u32,
+    change_tick: u32,
+}
+
+impl<'a, Trait: ?Sized + TraitQuery> Iterator for WriteTableTraitsIter<'a, Trait> {
+    type Item = Mut<'a, Trait>;
+    fn next(&mut self) -> Option<Self::Item> {
+        // Iterate the remaining table components that are registered,
+        // until we find one that exists in the table.
+        let (column, meta) = unsafe { zip_exact(&mut self.components, &mut self.meta) }
+            .find_map(|(&component, meta)| self.table.get_column(component).zip(Some(meta)))?;
+        let ptr = unsafe {
+            column
+                .get_data_ptr()
+                .byte_add(self.table_row * meta.size_bytes)
+        };
+        // SAFETY: The instance of `WriteTraits` that created this iterator
+        // has exclusive access to all table components registered with the trait.
+        //
+        // Since `self.table_row` is guaranteed to be unique, we know that other instances
+        // of `WriteTableTraitsIter` will not conflict with this pointer.
+        let ptr = unsafe { ptr.assert_unique() };
+        let trait_object = unsafe { meta.dyn_ctor.cast_mut(ptr) };
+        // SAFETY: We have exclusive access to the component, so by extension
+        // we have exclusive access to the corresponding `ComponentTicks`.
+        let component_ticks = unsafe { column.get_ticks_unchecked(self.table_row).deref_mut() };
+        Some(Mut {
+            value: trait_object,
+            ticks: Ticks {
+                component_ticks,
+                last_change_tick: self.last_change_tick,
+                change_tick: self.change_tick,
+            },
+        })
+    }
+}
+
+#[doc(hidden)]
+pub struct WriteSparseTraitsIter<'a, Trait: ?Sized> {
+    // SAFETY: These two iterators must have equal length.
+    components: std::slice::Iter<'a, ComponentId>,
+    meta: std::slice::Iter<'a, TraitImplMeta<Trait>>,
+    /// SAFETY: Given the same trait type and same archetype,
+    /// no two instances of this struct may have the same `entity`.
+    entity: Entity,
+    sparse_sets: &'a SparseSets,
+    last_change_tick: u32,
+    change_tick: u32,
+}
+
+impl<'a, Trait: ?Sized + TraitQuery> Iterator for WriteSparseTraitsIter<'a, Trait> {
+    type Item = Mut<'a, Trait>;
+    fn next(&mut self) -> Option<Self::Item> {
+        // Iterate the remaining sparse set components we have registered,
+        // until we find one that exists in the archetype.
+        let ((ptr, component_ticks), meta) =
+            unsafe { zip_exact(&mut self.components, &mut self.meta) }.find_map(
+                |(&component, meta)| {
+                    self.sparse_sets
+                        .get(component)
+                        .and_then(|set| set.get_with_ticks(self.entity))
+                        .zip(Some(meta))
+                },
+            )?;
+
+        // SAFETY: The instance of `WriteTraits` that created this iterator
+        // has exclusive access to all sparse set components registered with the trait.
+        //
+        // Since `self.entity` is guaranteed to be unique, we know that other instances
+        // of `WriteSparseTraitsIter` will not conflict with this pointer.
+        let ptr = unsafe { ptr.assert_unique() };
+        let trait_object = unsafe { meta.dyn_ctor.cast_mut(ptr) };
+        // SAFETY: We have exclusive access to the component, so by extension
+        // we have exclusive access to the corresponding `ComponentTicks`.
+        let component_ticks = unsafe { component_ticks.deref_mut() };
+
+        Some(Mut {
+            value: trait_object,
+            ticks: Ticks {
+                component_ticks,
+                last_change_tick: self.last_change_tick,
+                change_tick: self.change_tick,
+            },
+        })
+    }
+}
+
+impl<'w, Trait: ?Sized + TraitQuery> WriteTraits<'w, Trait> {
+    /// Returns an iterator over the components implementing `Trait` for the current entity.
+    pub fn iter(&self) -> CombinedReadTraitsIter<'_, Trait> {
+        self.into_iter()
+    }
+    /// Returns a mutable iterator over the components implementing `Trait` for the current entity.
+    pub fn iter_mut(&mut self) -> CombinedWriteTraitsIter<'_, Trait> {
+        self.into_iter()
+    }
+}
+
+impl<'w, Trait: ?Sized + TraitQuery> IntoIterator for WriteTraits<'w, Trait> {
+    type Item = Mut<'w, Trait>;
+    type IntoIter = CombinedWriteTraitsIter<'w, Trait>;
+    #[inline]
+    fn into_iter(self) -> Self::IntoIter {
+        let table = WriteTableTraitsIter {
+            components: self.registry.table_components.iter(),
+            meta: self.registry.table_meta.iter(),
+            table: self.table,
+            table_row: self.table_row,
+            last_change_tick: self.last_change_tick,
+            change_tick: self.change_tick,
+        };
+        let sparse = WriteSparseTraitsIter {
+            components: self.registry.sparse_components.iter(),
+            meta: self.registry.sparse_meta.iter(),
+            entity: self.table.entities()[self.table_row],
+            sparse_sets: self.sparse_sets,
+            last_change_tick: self.last_change_tick,
+            change_tick: self.change_tick,
+        };
+        table.chain(sparse)
+    }
+}
+
+impl<'world, 'local, Trait: ?Sized + TraitQuery> IntoIterator
+    for &'local WriteTraits<'world, Trait>
+{
+    type Item = &'local Trait;
+    type IntoIter = CombinedReadTraitsIter<'local, Trait>;
+    #[inline]
+    fn into_iter(self) -> Self::IntoIter {
+        let table = ReadTableTraitsIter {
+            components: self.registry.table_components.iter(),
+            meta: self.registry.table_meta.iter(),
+            table: self.table,
+            table_row: self.table_row,
+        };
+        let sparse = ReadSparseTraitsIter {
+            components: self.registry.sparse_components.iter(),
+            meta: self.registry.sparse_meta.iter(),
+            entity: self.table.entities()[self.table_row],
+            sparse_sets: self.sparse_sets,
+        };
+        table.chain(sparse)
+    }
+}
+
+impl<'world, 'local, Trait: ?Sized + TraitQuery> IntoIterator
+    for &'local mut WriteTraits<'world, Trait>
+{
+    type Item = Mut<'local, Trait>;
+    type IntoIter = CombinedWriteTraitsIter<'local, Trait>;
+    #[inline]
+    fn into_iter(self) -> Self::IntoIter {
+        let table = WriteTableTraitsIter {
+            components: self.registry.table_components.iter(),
+            meta: self.registry.table_meta.iter(),
+            table: self.table,
+            table_row: self.table_row,
+            last_change_tick: self.last_change_tick,
+            change_tick: self.change_tick,
+        };
+        let sparse = WriteSparseTraitsIter {
+            components: self.registry.sparse_components.iter(),
+            meta: self.registry.sparse_meta.iter(),
+            entity: self.table.entities()[self.table_row],
+            sparse_sets: self.sparse_sets,
+            last_change_tick: self.last_change_tick,
+            change_tick: self.change_tick,
+        };
+        table.chain(sparse)
+    }
+}
+
+#[doc(hidden)]
+pub struct WriteAllTraitsFetch<'w, Trait: ?Sized + TraitQuery> {
+    pub(crate) registry: &'w TraitImplRegistry<Trait>,
+    pub(crate) table: Option<&'w Table>,
+    pub(crate) sparse_sets: &'w SparseSets,
+    pub(crate) last_change_tick: u32,
+    pub(crate) change_tick: u32,
+}
